Have run with valgrind. No problems except
a couple of small mem leaks caused by tests +
a mem leak caused by PQ not freeing elements properly.
Mario pointed out that I am wasting memory by having pointers i my element anyway.
SO next step is to fix that - but best to commit first.

+++ include/pop_graph/file_reader.h
@@ -14,10 +14,10 @@ int load_fasta_data_from_filename_into_graph_of_specific_person_or_pop(char* fil
 int load_fastq_data_from_filename_into_graph_of_specific_person_or_pop(char* filename, long long * count_kmers, long long * bad_reads,  char quality_cut_off, int max_read_length, dBGraph* db_graph,
 								       EdgeArrayType type, int index);
 
-int load_population_as_fasta(char* filename,  long long* count_kmers, long long* bad_reads, dBGraph* db_graph);
+long long load_population_as_fasta(char* filename,  long long* count_kmers, long long* bad_reads, dBGraph* db_graph);
 int load_all_fasta_for_given_person_given_filename_of_file_listing_their_fasta_files(char* f_name, long long* count_kmers, long long* bad_reads, dBGraph* db_graph, int index);
 
-int load_population_as_fastq(char* filename,  long long* count_kmers, long long* bad_reads, char quality_cutoff, dBGraph* db_graph);
+long long load_population_as_fastq(char* filename,  long long* count_kmers, long long* bad_reads, char quality_cutoff, dBGraph* db_graph);
 int load_all_fastq_for_given_person_given_filename_of_file_listing_their_fastq_files(char* f_name, long long* count_kmers, long long* bad_reads, char quality_cutoff, dBGraph* db_graph, int index);
